\name{metEPF}
\alias{metEPF}
\title{Energy Pattern Factor Toolbox}
\usage{
metEPF(df=dataframe,start="chr ",end="chr")
}
\description{
    df:     df is a univariate dataframe of Ground station wind speed(m/s),
    start:  start is the intial time date string character,mdy e.g start="01-01-2012",
    end:    end is given as final time date string character, mdy e.g end="12-31-2013"
}
\examples{
metEPF(df,start,end)
}
